Repack

+++ uplift-json/src/main/java/com/github/kjetilv/uplift/json/io/StringSink.java
@@ -24,7 +24,7 @@ record StringSink(StringBuilder sb) implements Sink {
     }
 
     @Override
-    public int length() {
+    public long length() {
         return sb.length();
     }
 }
